+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## master (unreleased)
 
+## 0.32.0 (2025-04-05)
+
 * [#328](https://github.com/clojure-emacs/orchard/pull/328): Inspector: display identity hashcode for Java objects.
 * [#329](https://github.com/clojure-emacs/orchard/pull/329): Inspector: add analytics.
 * [#331](https://github.com/clojure-emacs/orchard/pull/331): Inspector: add table view mode.